// src/app/model/enums.ts
export enum EstadosDocumentoCargadoEnum {
    no_cargado = "No cargado",
    Cargado = "Cargado"
}

export enum UsuarioSesionEnum {
    incapacidad = "incapacidad"
}

export enum EstadosDocumentoEnum {
    SIN_VALIDAR = 'SIN_VALIDAR',
    APROBADO = 'APROBADO',
    RECHAZADO = 'RECHAZADO'
}

export enum EstadosRadicadoEnum {
    CPT = 'CPT',
    CAP = 'CAP',
    PEN = 'PEN',
    APR = 'APR',
    REH = 'REH',
    ANU = 'ANU',
    DEV = 'DEV',
    RAT = 'RAT',
    RNN = 'RNN',
    INR = 'INR',
    INP = 'INP',
    RIN = 'RIN',
    RRI = 'RRI'
}

export enum EstadosObservacionEnum {
    EN_CAPTURA = 'EN CAPTURA',
    EN_TRANSCRIPCION = 'EN TRANSCRIPCION',
    EN_INVESTIGACION = 'EN INVESTIGACION',
    DEVUELTA = 'DEVUELTA'
}